backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.1
pyodbc>=5.0.0
python-dotenv>=1.0.0
colorama>=0.4.6